// C0920G1-CarInsurance-FE/src/app/checkout/model/billDTO.ts
export class BillDTO {
  customer_name: string;
  customer_phone: string;
  customer_email: string;
  car_bks: string;
  car_year_manufacturing:string;
  car_manufacturing:string;
  contract_id: number;
  contract_start_date: string;
  contract_end_date: string
  contract_duration :string;
  product_name: string;
  product_type_name: string;
  product_price: string;
  employee_name: string;
  total_money: number;
  status_approval : string;
  payment:string;
  payment_type :string;
  status_pay:string;
  contract_detail_id;
}